Split ParticleID collection into smaller collections
Having one large ParticleID collection made sense historically before
EDM4hep reversed the direction of the ReconstructedParticle ->
ParticleID relation. Since now ParticleIDs point to the
ReconstructedParticles it's easier to handle downstream (and especially
in FCCAnalyses) if there are several smaller and more specific
ParticleID collections for each of the input collections that are
converted.